+++ entity/src/user_projects.rs
@@ -6,9 +6,9 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "user_projects")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub user_id: i32,
+    pub user_id: i64,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub project_id: i32,
+    pub project_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
